Fix up the testing to get everything passing.
There are still a few tests being skipped, but it's something I'll need to look more into before I can make a diagnosis.
We currently have some basic functionality in place which can be used to start with the next section of work.

+++ utilities.go
@@ -43,7 +43,7 @@ type SysInfo struct {
 	} `json:"system"`
 	Time struct {
 		Current      Timestamp `json:"current"`
-		Uptime       int64     `json:"uptime"`
+		Uptime       float64   `json:"uptime"`
 		Timezone     string    `json:"timezone"`
 		TimezoneName string    `json:"timezoneName"`
 	} `json:"time"`
